Remove `DataTaskDelegate` and reuse `StreamingDelegate` (but adapted "no Combine")
